Implement `ViewMotionValue` runtime independent of compose
Provide a Java-like interface (intentionally using callbacks instead of
flows for a zero-requirements API)
Use the previously extracted `Computations` to compute the output.

This is only intended to be used where using Compose's Snapshot state is
not an option. All other clients should use the regular MotionValue (with SnapshotViewBinding if needed)

Flag: EXEMPT not used yet
Bug: 389637975
Test: atest mechanics_test
Change-Id: I3c4aed46f58ab2974deaff5763ed5c120ee2e0de