Add flags for consolidated weather cards.
Added separate flags for data and layout. No need to block all the work
on the consolidated layout by the work on semantic data.

The work on the layout is low risk. The plan is to implement it in a way
that there is no visible effect unless changes are made upstream in the
data pipeline. Therefore setting the flag as BUGFIX.

Flag: com.android.systemui.shared.smartspace_semantic_weather_data
Flag: com.android.systemui.shared.smartspace_consolidated_weather_layout
Bug: 416745891
Test: N/A. Just adding flags.
Change-Id: Ia3e8e0b1385db5bfc4defbc0457b567cd1786a22