[3/n] Update DisplayManager to persist display connection preference
The new dialog has a checkbox to allow users to save their preferred
connection type for each individual display. To perist this data so it
can be written & read, we are using PersistentDataStore and storing it
as a key value pair in DisplayState, with keys being a displays unique
id.

Bug: 413620089
Test: ConnectedDisplayInteractorTest
Test: DisplayRepositoryTest
Flag: com.android.window.flags.enable_updated_display_connection_dialog
Change-Id: I4fde2c1b23e9aa98de5bb65c8db1ebca913dc566